Design play

+++ send.php
@@ -8,40 +8,66 @@ if($_POST){
     $msg = '';
     $mail = new PHPMailer;
 
-    $image_data = $_POST['img'];
-    $image_data = str_replace('data:image/png;base64,', '', $image_data);
-    $image_data = str_replace(' ', '+', $image_data);
-    $data = base64_decode($image_data);
-    $filename = 'image.png';
-    file_put_contents($filename, $data);
-    //$uploadfile = substr($image_data, strpos($image_data, ","));
-    //$uploadfile = str_replace(" ", "+", substr($image_data, strpos($image_data, ",")+1));
-    //$uploadfile = str_replace(' ','+',$image_data);
-    //$uploadfile2 = base64_decode($uploadfile);
+    if(array_key_exists('name', $_POST))
+    {
+        $name = $_POST['name'];
+    }
+    else {
+        $name = 'Unknown';
+    }
 
-//  $fp = fopen('test.png', 'w');
-//  fwrite($fp, $uploadfile);
-//  fclose($fp);
+    if(array_key_exists('work', $_POST))
+    {
+        $work = $_POST['work'];
+    }
+    else {
+        $work = 'Unknown';
+    }
+
+    if(array_key_exists('contact', $_POST))
+    {
+        $contact = $_POST['contact'];
+    }
+    else {
+        $contact = 'Unknown';
+    }
+
+    if(array_key_exists('contact', $_POST))
+    {
+        $imgDataMCU = $_POST['imgData'];
+    }
+    else {
+        $imgDataMCU = 'Blank';
+    }
+    
 
     $mail->setFrom('info@ipanel.com', 'PixelDrawer');
     $mail->addAddress('coworking.shostka@gmail.com', 'Коворкинг Шостка');
 
     $mail->isHTML(true);
 
-    $mail->Subject = $_POST['name'];
-    $mail->Body = "Name: {$_POST['name']}<br> Work: {$_POST['work']} <br> $uploadfile";
-    //$mail-> AddStringAttachment($uploadfile, 'Image.png');
-    // addStringEmbeddedImage
-    // data from url?
+    $mail->Subject = $name;
+    $mail->Body = "Name: $name <br> Work: $work <br> Contact: $contact <br> Img Data for MCU: $imgDataMCU";
+    
+    if(array_key_exists('img', $_POST))
+    {
+    $image_data = $_POST['img'];
+    $image_data = str_replace('data:image/png;base64,', '', $image_data);
+    $image_data = str_replace(' ', '+', $image_data);
+    $data = base64_decode($image_data);
+    $filename = 'image.png';
+    file_put_contents($filename, $data);
     $mail->addAttachment($filename, 'MyImage');
+    }   
+
     if (!$mail->send()) {
-        $msg .= "Mailer Error: " . $mail->ErrorInfo;
+        // $msg .= "Mailer Error: " . $mail->ErrorInfo;
         $answer = '1';
     } else {
-        $msg .= "Message sent!";
+        // $msg .= "Message sent!";
         $answer = '0';
     }
-    die( $msg );
+    die( $answer );
 }
 ?>
